Implement Schell's Table
Ran into a few issues while testing. I think the way hashing works
is that every time the program runs a new seed is used causing the
hash from the previous compilations to not match up with the current.

I don't know why it took me so long to switch to just using rank as
the key, but as it stands now it works. Need to optimize the
efficiency as well as the naming. Some code clutter could be cleaned
up as well.